@@ -113,10 +113,10 @@ class ratpanel implements DeployInterface
         $url = $this->url . '/api' . $path;
         $body = $method == 'GET' ? null : json_encode($params);
         $sign = $this->signRequest($method, $url, $body, $this->id, $this->token);
-        $response = curl_client($url, $body, null, null, [
-            'Content-Type: application/json',
-            'X-Timestamp: ' . $sign['timestamp'],
-            'Authorization: HMAC-SHA256 Credential=' . $sign['id'] . ', Signature=' . $sign['signature']
+        $response = http_request($url, $body, null, null, [
+            'Content-Type' => 'application/json',
+            'X-Timestamp' => $sign['timestamp'],
+            'Authorization' => 'HMAC-SHA256 Credential=' . $sign['id'] . ', Signature=' . $sign['signature']
         ], $this->proxy, $method);
         return $response['body'];
     }
@@ -130,7 +130,7 @@ class ratpanel implements DeployInterface
 
         // 规范化路径
         $canonicalPath = $path;
-        if (strpos($path, '/api') !== 0) {
+        if (!str_starts_with($path, '/api')) {
             $apiPos = strpos($path, '/api');
             if ($apiPos !== false) {
                 $canonicalPath = substr($path, $apiPos);
