@@ -27,7 +27,7 @@ class Ucloud
         $param = array_merge($param, $params);
         $param['Signature'] = $this->ucloudSignature($param);
         $ua = sprintf("PHP/%s PHP-SDK/%s", phpversion(), self::VERSION);
-        $response = get_curl($this->ApiUrl, json_encode($param), 0, 0, 0, $ua, 0, ['Content-Type: application/json']);
+        $response = get_curl($this->ApiUrl, json_encode($param), 0, 0, $ua, 0, ['Content-Type' => 'application/json']);
         $result = json_decode($response, true);
         if (isset($result['RetCode']) && $result['RetCode'] == 0) {
             return $result;
