@@ -42,13 +42,13 @@ class gcore implements DeployInterface
     private function request($path, $params = null, $method = null)
     {
         $url = $this->url . $path;
-        $headers = ['Authorization: APIKey ' . $this->apikey];
+        $headers = ['Authorization' => 'APIKey ' . $this->apikey];
         $body = null;
         if ($params) {
-            $headers[] = 'Content-Type: application/json';
+            $headers['Content-Type'] = 'application/json';
             $body = json_encode($params);
         }
-        $response = curl_client($url, $body, null, null, $headers, $this->proxy, $method);
+        $response = http_request($url, $body, null, null, $headers, $this->proxy, $method);
         $result = json_decode($response['body'], true);
         if ($response['code'] >= 200 && $response['code'] < 300) {
             return $result;
